k8s/requirements.txt
google-cloud-bigtable==2.21.0
functions-framework==3.4.0